refactor rendering of the panel
