+++ cosmic-panel-bin/src/space/render.rs
@@ -3,19 +3,23 @@ use std::time::Duration;
 use super::{panel_space::MyRenderElements, PanelSpace};
 use cctk::wayland_client::{Proxy, QueueHandle};
 use cosmic_panel_config::PanelAnchor;
+use image::RgbaImage;
 use itertools::Itertools;
 use sctk::shell::WaylandSurface;
 use smithay::{
-    backend::renderer::{
-        damage::OutputDamageTracker,
-        element::{
-            memory::MemoryRenderBufferRenderElement,
-            surface::{render_elements_from_surface_tree, WaylandSurfaceRenderElement},
+    backend::{
+        allocator::Fourcc,
+        renderer::{
+            damage::OutputDamageTracker,
+            element::{
+                memory::{MemoryRenderBuffer, MemoryRenderBufferRenderElement},
+                surface::{render_elements_from_surface_tree, WaylandSurfaceRenderElement},
+            },
+            gles::GlesRenderer,
+            Bind, Frame, Renderer, Unbind,
         },
-        gles::GlesRenderer,
-        Bind, Frame, Renderer, Unbind,
     },
-    utils::{Logical, Point, Rectangle, Size},
+    utils::{Logical, Point, Rectangle, Transform},
 };
 use xdg_shell_wrapper::{shared_state::GlobalState, space::WrapperSpace};
 
@@ -37,6 +41,9 @@ impl PanelSpace {
         }
 
         if self.is_dirty && self.has_frame {
+            if let Err(err) = self.render_panel(bg_color) {
+                tracing::error!(?err, "Error rendering the panel.");
+            }
             let my_renderer = match self.damage_tracked_renderer.as_mut() {
                 Some(r) => r,
                 None => return Ok(()),
@@ -116,7 +123,7 @@ impl PanelSpace {
                 // this is a workaround
                 if !self.first_draw {
                     if let Some(buff) = self.buffer.as_mut() {
-                        let mut render_context = buff.render();
+                        let render_context = buff.render();
                         let margin_offset = match self.config.anchor {
                             PanelAnchor::Top | PanelAnchor::Left => {
                                 self.config.get_effective_anchor_gap() as f64
@@ -124,9 +131,7 @@ impl PanelSpace {
                             PanelAnchor::Bottom | PanelAnchor::Right => 0.0,
                         };
 
-                        let (panel_size, loc) = if let Some(animate_state) =
-                            self.animate_state.as_ref()
-                        {
+                        let loc = if let Some(animate_state) = self.animate_state.as_ref() {
                             let actual_length = if self.config.is_horizontal() {
                                 self.actual_size.w
                             } else {
@@ -149,30 +154,13 @@ impl PanelSpace {
                                 }
                                 PanelAnchor::Bottom | PanelAnchor::Right => 0.0,
                             };
-                            let crosswise_length = if self.config.is_horizontal() {
-                                self.actual_size.h
+
+                            let (x, y) = if self.config.is_horizontal() {
+                                (lengthwise_pos, crosswise_pos)
                             } else {
-                                self.actual_size.w
+                                (crosswise_pos, lengthwise_pos)
                             };
-                            let (x, y, width, height) = if self.config.is_horizontal() {
-                                (
-                                    lengthwise_pos,
-                                    crosswise_pos,
-                                    container_length,
-                                    crosswise_length,
-                                )
-                            } else {
-                                (
-                                    crosswise_pos,
-                                    lengthwise_pos,
-                                    crosswise_length,
-                                    container_length,
-                                )
-                            };
-                            (
-                                Size::<i32, Logical>::from((width, height)),
-                                Point::<f64, Logical>::from((x, y)),
-                            )
+                            Point::<f64, Logical>::from((x, y))
                         } else if is_dock {
                             let loc: Point<f64, Logical> = if self.config.is_horizontal() {
                                 (
@@ -187,7 +175,7 @@ impl PanelSpace {
                             }
                             .into();
 
-                            (self.actual_size, loc)
+                            loc
                         } else {
                             let loc: Point<f64, Logical> = if self.config.is_horizontal() {
                                 (0.0, margin_offset)
@@ -196,21 +184,9 @@ impl PanelSpace {
                             }
                             .into();
 
-                            (self.dimensions, loc)
+                            loc
                         };
-                        let scaled_panel_size =
-                            panel_size.to_f64().to_physical(self.scale).to_i32_round();
 
-                        let _ = render_context.draw(|_| {
-                            if self.buffer_changed {
-                                Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(
-                                    Point::default(),
-                                    (scaled_panel_size.w, scaled_panel_size.h),
-                                )])
-                            } else {
-                                Result::<_, ()>::Ok(Default::default())
-                            }
-                        });
                         self.buffer_changed = false;
 
                         drop(render_context);
@@ -252,6 +228,8 @@ impl PanelSpace {
 
                 self.is_dirty = false;
                 self.has_frame = false;
+
+                // TODO clear the stencil buffer
             }
         }
 
@@ -297,6 +275,130 @@ impl PanelSpace {
         }
         renderer.unbind()?;
         self.first_draw = false;
+
+        Ok(())
+    }
+
+    fn render_panel(&mut self, bg_color: [f32; 4]) -> anyhow::Result<()> {
+        if !self.buffer_changed {
+            return Ok(());
+        }
+
+        let gap = self.gap();
+        let border_radius = self.border_radius();
+        let mut panel_size = self.actual_size;
+        let container_length = self.container_length;
+
+        if self.config.is_horizontal() {
+            panel_size.w = container_length as i32;
+        } else {
+            panel_size.h = container_length as i32;
+        }
+        let panel_size = panel_size.to_f64().to_physical(self.scale).to_i32_round();
+
+        let bg_color: [u8; 4] = bg_color
+            .iter()
+            .map(|c| ((c * 255.0) as u8).clamp(0, 255))
+            .collect_vec()
+            .try_into()
+            .unwrap();
+
+        let radius = (border_radius as f64 * self.scale).round() as u32;
+        let radius = radius
+            .min(panel_size.w as u32 / 2)
+            .min(panel_size.h as u32 / 2);
+        let mut calculated_corner_image = None;
+        // TODO use 2 MemoryRenderBuffer for sides, and 1 single pixel buffer for center
+        let mut buff = MemoryRenderBuffer::new(
+            Fourcc::Abgr8888,
+            (panel_size.w, panel_size.h),
+            1,
+            Transform::Normal,
+            None,
+        );
+
+        let mut render_context = buff.render();
+        let _ = render_context.draw(|buffer| {
+            buffer.chunks_exact_mut(4).for_each(|chunk| {
+                chunk.copy_from_slice(&bg_color);
+            });
+
+            // early return if no radius
+            if radius == 0 {
+                return Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(
+                    Point::default(),
+                    (panel_size.w, panel_size.h),
+                )]);
+            }
+            let drawn_radius = 128;
+            let drawn_radius2 = drawn_radius as f64 * drawn_radius as f64;
+            let grid = (0..((drawn_radius + 1) * (drawn_radius + 1)))
+                .into_iter()
+                .map(|i| {
+                    let (x, y) = (i as u32 % (drawn_radius + 1), i as u32 / (drawn_radius + 1));
+                    drawn_radius2 - (x as f64 * x as f64 + y as f64 * y as f64)
+                })
+                .collect_vec();
+
+            let empty = [0, 0, 0, 0];
+
+            let mut corner_image = RgbaImage::new(drawn_radius, drawn_radius);
+            for i in 0..(drawn_radius * drawn_radius) {
+                let (x, y) = (i as u32 / drawn_radius, i as u32 % drawn_radius);
+                let bottom_left = grid[(y * (drawn_radius + 1) + x) as usize];
+                let bottom_right = grid[(y * (drawn_radius + 1) + x + 1) as usize];
+                let top_left = grid[((y + 1) * (drawn_radius + 1) + x) as usize];
+                let top_right = grid[((y + 1) * (drawn_radius + 1) + x + 1) as usize];
+                let color = if bottom_left >= 0.0
+                    && bottom_right >= 0.0
+                    && top_left >= 0.0
+                    && top_right >= 0.0
+                {
+                    bg_color.clone()
+                } else {
+                    empty
+                };
+                corner_image.put_pixel(x, y, image::Rgba(color));
+            }
+            let corner_image = image::imageops::resize(
+                &corner_image,
+                radius as u32,
+                radius as u32,
+                image::imageops::FilterType::CatmullRom,
+            );
+
+            for (i, color) in corner_image.pixels().enumerate() {
+                let (x, y) = (i as u32 % radius, i as u32 / radius);
+                let top_left = (radius - 1 - x, radius - 1 - y);
+                let top_right = (panel_size.w as u32 - radius + x, radius - 1 - y);
+                let bottom_left = (radius - 1 - x, panel_size.h as u32 - radius + y);
+                let bottom_right = (
+                    panel_size.w as u32 - radius + x,
+                    panel_size.h as u32 - radius + y,
+                );
+                for (c_x, c_y) in match (self.config.anchor, gap > 0) {
+                    (PanelAnchor::Left, false) => vec![top_right, bottom_right],
+                    (PanelAnchor::Right, false) => vec![top_left, bottom_left],
+                    (PanelAnchor::Top, false) => vec![bottom_left, bottom_right],
+                    (PanelAnchor::Bottom, false) => vec![top_left, top_right],
+                    _ => vec![top_left, top_right, bottom_left, bottom_right],
+                } {
+                    let b_i = (c_y * panel_size.w as u32 + c_x) as usize * 4;
+                    let c = buffer.get_mut(b_i..b_i + 4).unwrap();
+                    c.copy_from_slice(&color.0);
+                }
+            }
+
+            calculated_corner_image = Some(corner_image);
+            // Return the whole buffer as damage
+            Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(
+                Point::default(),
+                (panel_size.w, panel_size.h),
+            )])
+        });
+        drop(render_context);
+        self.buffer = Some(buff);
+        self.buffer_changed = true;
 
         Ok(())
     }
